cf/requirements.txt
Flask<3.0
google-cloud-pubsub==2.20.1